Create youtube
In this code, we start by asking the user to enter the YouTube video URL using input() function. We then create a YouTube object with the URL of the video.

We print some information about the video, including the video title, thumbnail URL, and duration. We then select the highest resolution stream available for the video using get_highest_resolution().

Finally, we download the video using download() and print a message to indicate that the download is completed.

Note that this is a simple code and doesn't include any error handling or user input validation. You may want to add some error handling code to handle cases where the video is not available or the user enters an invalid URL.